### 2016-04-11 #35 @Dahaniel
The Database of Odorant Responses (DoOR.data)
- 1. What does this package do? (explain in 50 words or less)

  The data package of the DoOR project. The data package provides all published odor responses recorded from olfactory sensory neurons of the fruit fly _Drosophila_.
- 1. Paste the full DESCRIPTION file inside a code block below.

```
Package: DoOR.data
Title: Integrating Heterogeneous Odorant Response Data into a Common Response
    Model: A DoOR to the Complete Olfactome
Version: 2.0.0
    person("Daniel", "Münch", email = "daniel.muench@uni-konstanz.de", role = c("aut", "cre")),
    person("C. Giovanni", "Galizia", email = "giovanni.galizia@uni-konstanz.de", role = "aut"),
    person("Shouwen", "Ma", role = "aut"),
    person("Martin", "Strauch", role = "aut"),
    person("Anja", "Nissler", role = "aut"))
Description: This is a data package providing Drosophila odorant response data for
    DoOR.functions. See http://dx.doi.org/10.1093/chemse/bjq042 for the original
    publication and http://dx.doi.org/10.1038/srep21841 regarding version 2.0.
License: CC BY-SA 4.0
    utils
    DoOR.functions (>= 2.0.0)
    Dahaniel/DoOR.functions@v2.0.0

```
- 3. URL for the package (the development repository, not a stylized html page)

https://github.com/Dahaniel/DoOR.data
- 4. What data source(s) does it work with (if applicable)?

The package itself provides the data either extracted from publications or provided by the publication authors.
- 5. Who is the target audience?

Scientists working on olfactory coding, modelers and physiologists.
- 6. Are there other R packages that accomplish the same thing? If so, what is different about yours?

No, to our knowledge DoOR is the only project providing this kind of data and using this approach to integrate heterogeneous data.
- 7. Check the box next to each policy below, confirming that you agree. These are mandatory.
  - [x] This package does not violate the Terms of Service of any service it interacts with.
  - [x] The repository has continuous integration with Travis CI and/or another service
  - [ ] The package contains a vignette
  - [x] The package contains a reasonably complete README with `devtools` install instructions
  - [ ] The package contains unit tests
  - [x] The package only exports functions to the NAMESPACE that are intended for end users
- 8. Do you agree to follow the [rOpenSci packaging guidelines](https://github.com/ropensci/packaging_guide)? These aren't mandatory, but we strongly suggest you follow them. If you disagree with anything, please explain.
  - [ ] Are there any package dependencies not on CRAN? - _no_
  - [x] Does the package have a CRAN accepted license?
  - [ ] Did `devtools::check()` produce any errors or warnings? If so paste them below. - _no_
- 1. Please add explanations below for any exceptions to the above:
  - The package does not contain a vignette as it only provides data. All the functionality is comprehensively explained in the vignettes in the corresponding data package (see other onboarding).
  - I think unit tests do not make much sense here as the package solely provides data.
  - Travis has problems as the DoOR.data package suggests the DoOR.functions package. Adding DoOR.functions to "Remotes:" did not help as it depends on DoOR.data...
- 1. If this is a resubmission following rejection, please explain the change in circumstances.
