### 2016-04-17 #35 @noamross
Reviewers: @RemkoDuursma
Due date: 2016-05-09
