### 2016-05-03 #35 @RemkoDuursma
This package includes data on olfactory responses of Drosophila. I do now know anything about this kind of data. I have reviewed this package only in terms of the R code, functionality, interface, and documentation.

The package only includes data as a set of dataframes, and no other functions. All data is loaded into the current workspace with `loadData()`. I find this choice of data porting somewhat unfortunate, as it simply loads all included datasets into the current workspace, potentially causing conflicts and certainly causing a messy workspace. But perhaps there is a good reason to do this, as I am not sure how these kind of data are used in analyses.

The 'standard' way of loading data is also available, though it seems not to be documented. For example we can do, `data("Or94b")` to load that particular dataset into the current workspace. The help page for `loadData` should show examples of this (and it should be renames `load_all_DoOr_data`, or similarly descriptive, see below).

The choice of data storage, in binary format (`Rdata` in data/) is fine in terms of storage space but prevents meaningful version control on the data. It is not clear to me (again due to my ignorance) if the underlying datasets are prone to change over time, or whether this is an unchanging database. If there is change to be expected within datasets, a different format would be preferable where incremental changes can be logged with version control.

I found the documentation to be extremely concise. For example the help file for `?data.format` only states "DoOr response data template", and all other help files are similarly brief. Surely more information is needed in order to use the package effectively (it would not be sufficient in my mind if this information can be found elsewhere).

The function names should follow protocol. Snake case is preferred, and the package should avoid datasets with very generic names (`response.matrix`,`dataset.info`,`data.format`, etc.). I would certainly avoid the use of `.` in names of functions to avoid possible S3 class dispatches.

In summary the package contains data only (I cannot judge whether these data are useful or broadly applicable), but some should be renamed, and documentation improved.
