### 2016-05-08 #35 @sckott
hey there, for full openness - the bot commented above that a review wasn't in - but of course it is in. The offending line is here https://github.com/ropenscilabs/heythere/blob/master/helpers.rb#L41 - so we obviously need a more sophisticated way to tell if a review is submitted than how long a comment is.
