### 2016-05-08 #35 @RemkoDuursma
I can also conclude that my review is the shortest yet! It should have been
at least 3000 words :)

-------------------------------------------------
Remko Duursma
Senior Lecturer

Hawkesbury Institute for the Environment
Western Sydney University
Hawkesbury Campus, Richmond

Office: +61(0)2 4570 1806
Mobile: +61 (0)4 22 096908
www.remkoduursma.com

On Sun, May 8, 2016 at 4:22 PM, Scott Chamberlain <notifications@github.com>
wrote:

> hey there, for full openness - the bot commented above that a review
> wasn't in - but of course it is in. The offending line is here
> https://github.com/ropenscilabs/heythere/blob/master/helpers.rb#L41 - so
> we obviously need a more sophisticated way to tell if a review is submitted
> than how long a comment is.
>
