### 2016-05-09 #35 @sckott
@RemkoDuursma well it was a data only a package, so short is okay in this case.
