### 2016-05-09 #35 @sckott
@Dahaniel

* It would be nice to see better documentation of datasets here https://github.com/Dahaniel/DoOR.data/blob/master/R/DoOR.data-package.R - that is, describe what each column name is and what data the column has, and any units if applicable
* Also if datasets were from different sources
* Here https://github.com/Dahaniel/DoOR.data/blob/master/R/loadData.R#L7 - I think there is a typical notation for email, like `\email{foo@bar.com}`
