### 2016-05-11 #35 @Dahaniel
First of all thank you very much for taking the time to review our package.

I agree that the datasets need more documentation and I will look into that. Regarding the other points:

* **loading to current workspace**: I know this might be a problem and I am looking into possibilities to change that. We are doing that because the user should be able to easily and directly modify the loaded datasets (e.g. add or exclude rows or columns), and then rerun analysis tools (provided by `DoOR.functions`). Separate environments were not easily visible to the user. Any advice would be appreciated.

* **binary format** We initially stored the data as csv for exactly the reason of being able to track changes. We switched to binary to save some space. I would happily switch back to .csv. With .csv, could there be problems arising across operating systems e.g. regarding encoding of line-breaks?

* **function names** We tried to follow the [Google R style-guide](https://google.github.io/styleguide/Rguide.xml) i.e. camel code for function names and variable names separated by dots. In general I refrained from changing names as people are used to the established names. But I guess we could adapt all names to ROpenSci guide lines and make an announcement on the webpage and maybe print an message upon loading the package.

@sckott regarding email annotations: are you referring to the double @ sign? I think I copied that from somewhere but I just tested it and roxygen seems to convert both to proper email addresses in the help file.
