### 2016-05-11 #35 @noamross
1) While snake case functions are preferred, for a package with existing users and consistent style other styles are fine.  If you wish to change function names to make them less generic and avoid `.`, I suggest using both in the next version of the package, with documentation/messages that the old versions are deprecated, and then removing the deprecated names in a future update.

2) In general, CSV should be interoperable between systems. Interoperability issues most commonly arise when Excel is somehow involved, as it encodes line breaks inconsistently across systems/versions.

As to loading in the workspace, I'd await @RemkoDuursma's review of DoOR.functions for his thoughts on the relationship between the two.
