### 2016-08-10 #35 @sckott
@Dahaniel where are we at here?  Anything we can help with?
