### 2016-08-11 #35 @Dahaniel
@sckott sorry, I did not find the time to apply the changes yet.
I will change data back to csv, work on better documentation and explore ways to load data not to current workspace.
Regarding function names I like the idea of making a slow transition, having old and new names work for a while and prompting a message explaining that. Is there an easy way to implement different names for the same function? Like defining an alias or similar?
