### 2016-08-11 #35 @sckott
You can simply do e.g., https://github.com/ropensci/RNeXML/blob/master/R/nexml_read.R#L70 and then use `@aliases` tag https://github.com/ropensci/RNeXML/blob/master/R/nexml_read.R#L8 to make sure it's linked and exported
