### 2017-03-07 #35 @Dahaniel
Please see: https://github.com/ropensci/onboarding/issues/34#issuecomment-284765039
