### 2017-04-17 #35 @sckott
@Dahaniel

- R check reveals many problems: you need to Rbuildignore `.travis.yml`/`CHANGELOG.md` (but you should also have a `NEWS.md` or `NEWS` as that's more typical of R packages), `Undefined global functions or variables: data` - and more
- there' a note about compressing your data better
- you can still write minimal tests even though this is a data pkg, e.g. load data, and make sure data is of right dimensions, etc.

Some of R check output

```
   Found the following hidden files and directories:
     .travis.yml
   These were most likely included in error. See section ‘Package
   structure’ in the ‘Writing R Extensions’ manual.

   CRAN-pack does not know about
     .travis.yml
✔  checking for portable file names
✔  checking for sufficient/correct file permissions
✔  checking whether package ‘DoOR.data’ can be installed
✔  checking installed package size
✔  checking package directory
✔  checking DESCRIPTION meta-information
N  checking top-level files
   Non-standard file/directory found at top level:
     ‘CHANGELOG.md’
✔  checking for left-over files
✔  checking index information
✔  checking package subdirectories
✔  checking R files for non-ASCII characters
✔  checking R files for syntax errors
✔  checking whether the package can be loaded
✔  checking whether the package can be loaded with stated dependencies
✔  checking whether the package can be unloaded cleanly
✔  checking whether the namespace can be loaded with stated dependencies
✔  checking whether the namespace can be unloaded cleanly
✔  checking use of S3 registration
✔  checking dependencies in R code
✔  checking S3 generic/method consistency
✔  checking replacement functions
✔  checking foreign function calls
N  checking R code for possible problems
   loadData: no visible global function definition for ‘data’
   load_door_data: no visible global function definition for ‘data’
     data
   Consider adding
     importFrom("utils", "data")
   to your NAMESPACE file.

   Found the following calls to data() loading into the global environment:
   File ‘DoOR.data/R/load_door_data.R’:
     data(list = all.data$results[, 3])
   See section ‘Good practice’ in ‘?data’.
✔  checking Rd files
✔  checking Rd metadata
✔  checking Rd line widths
✔  checking Rd cross-references
✔  checking for missing documentation entries
✔  checking for code/documentation mismatches
✔  checking Rd \usage sections
✔  checking Rd contents
✔  checking for unstated dependencies in examples
✔  checking contents of ‘data’ directory
✔  checking data for non-ASCII characters
W  checking data for ASCII and uncompressed saves

     Note: significantly better compression could be obtained
           by using R CMD build --resave-data
                       old_size new_size compress
     door_AL_map.RData    514Kb    396Kb       xz
```
