### 2017-04-19 #35 @Dahaniel
Thanks @sckott.

> * R check reveals many problems: you need to Rbuildignore
.travis.yml/CHANGELOG.md (but you should also have a NEWS.md or NEWS as that's
more typical of R packages), Undefined global functions or variables: data - and
more

ok

> * there' a note about compressing your data better

This is as I went back to save the original data in csv as suggested by
@RemkoDuursma above (https://github.com/ropensci/onboarding/issues/35#issuecomment-216434323) to facilitated version control of the data. I would rather keep the data as csv.

> * you can still write minimal tests even though this is a data pkg, e.g. load
data, and make sure data is of right dimensions, etc.

ok
