### 2017-04-19 #35 @sckott
Okay on the csv - since you intend to send to CRAN, they may complain about that, but you can address that then
