### 2017-07-15 #35 @Dahaniel
see https://github.com/ropensci/onboarding/issues/34#issuecomment-315562774
