### 2017-08-25 #35 @Dahaniel
Done, please see https://github.com/ropensci/onboarding/issues/34#issuecomment-324934191
